feat : 웹 소캣 연결 시도

+++ android/moiroom/app/src/main/java/com/example/moiroom/data/ChatRoom.kt
@@ -3,8 +3,8 @@ package com.example.moiroom.data
 import java.time.Instant
 
 data class ChatRoom(
-    val chatRoomId: Int,
-    val memberId: Int,
+    val chatRoomId: Long,
+    val memberId: Long,
     val memberNickname: String,
     val profileImageUrl: String,
     val lastMessage: String,
